extract helpers methods  from test code
